add setup to dump_stage
