+++ bin/reth/src/dump_stage/hashing_storage.rs
@@ -1,13 +1,11 @@
 use crate::{
     db::DbTool,
     dirs::{DbPath, PlatformPath},
+    dump_stage::setup,
 };
 use eyre::Result;
-use reth_db::{
-    cursor::DbCursorRO, database::Database, table::TableImporter, tables, transaction::DbTx,
-};
+use reth_db::{database::Database, table::TableImporter, tables};
 use reth_provider::Transaction;
-use reth_staged_sync::utils::init::init_db;
 use reth_stages::{stages::StorageHashingStage, Stage, StageId, UnwindInput};
 use std::ops::DerefMut;
 use tracing::info;
@@ -19,25 +17,7 @@ pub(crate) async fn dump_hashing_storage_stage<DB: Database>(
     output_db: &PlatformPath<DbPath>,
     dry_run: bool,
 ) -> Result<()> {
-    assert!(from < to, "FROM block should be bigger than TO block.");
-
-    info!(target: "reth::cli", "Creating separate db at {}", output_db);
-
-    let output_db = init_db(output_db)?;
-
-    output_db.update(|tx| {
-        tx.import_table_with_range::<tables::BlockTransitionIndex, _>(
-            &db_tool.db.tx()?,
-            Some(from - 1),
-            to + 1,
-        )
-    })??;
-
-    // Find the latest block to unwind from
-    let (tip_block_number, _) = db_tool
-        .db
-        .view(|tx| tx.cursor_read::<tables::BlockTransitionIndex>()?.last())??
-        .expect("some");
+    let (output_db, tip_block_number) = setup::<DB>(from, to, output_db, db_tool)?;
 
     // Dry-run an unwind to FROM block, so we can get the PlainStorageState and
     // PlainAccountState safely. There might be some state dependency from an address
